Optimize Relation-to-Node assosication import
There happens to be cases where the osm data misses the nodes that are a
part of a relation. In such cases, the query that makes Relation-to-Node
assosciation fails to find the temp_point with osmid that is in the
power_relation_members table.

Overcome this issue by using a INNER JOIN so that only those points that
are in temp_points tables are consider for adding the relations.

Also, improve logging for the script.