Refine highlighting on Relations' clustergroup
Performance of hightlighting of the relation clusters on mouseover
has been very bad till this commit. The highlighting wouldn't be removed
from some clusters after mouseout event unless mouseover and mouseout
occur on that cluster again. That seem to have been so because of the
implementation. We try to add dynamic HTMLclasses to the clusters' icon
HTML template and re-render the whole cluster group which was causing
performance issues.

We now identify these clusters and add classes dynamically via HTML DOM
handling instead of re-rendering the cluster icons. This gave a pretty
good performance gain and better hightlighting of clusters on mouse
events.