Fix import script to not skip any powerline
Until before this commit, the script will skip the import of powerlines
that have some points that are not tagged with `power=*` because we
don't import points that are not tagged with `power=*` and there are
quite some powerlines that have points without the tag. So, we now use a
temporary points tables to lookup and import them all.

Also, update the scripts to not import duplicates.